requirements.txt
async-timeout==4.0.3
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-JWT-Extended==4.5.3
flask-redis==0.4.0
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.4
imageio==2.32.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
redis==5.0.1
Werkzeug==3.0.1
WTForms==3.1.1
